Extract github client extensions into module
