+++ lib/robot.js
@@ -1,9 +1,8 @@
 const {EventEmitter} = require('promise-events')
-const GitHubApi = require('github')
-const Bottleneck = require('bottleneck')
 const express = require('express')
 const Context = require('./context')
 const logger = require('./logger')
+const GitHubApi = require('./github')
 
 /**
  * The `robot` parameter available to apps
@@ -137,7 +136,7 @@ class Robot {
    * @private
    */
   async auth (id) {
-    let github
+    const github = new GitHubApi({debug: process.env.LOG_LEVEL === 'trace'})
 
     if (id) {
       const res = await this.cache.wrap(`app:${id}:token`, () => {
@@ -145,33 +144,13 @@ class Robot {
         return this.app.createToken(id)
       }, {ttl: 60 * 59}) // Cache for 1 minute less than GitHub expiry
 
-      github = new GitHubApi({debug: process.env.LOG_LEVEL === 'trace'})
       github.authenticate({type: 'token', token: res.data.token})
     } else {
-      github = await this.app.asApp()
+      github.authenticate((await this.app.asApp()).auth)
     }
 
-    return probotEnhancedClient(github)
+    return github
   }
-}
-
-function probotEnhancedClient (github) {
-  github = rateLimitedClient(github)
-
-  github.paginate = require('./paginate')
-
-  return github
-}
-
-// Hack client to only allow one request at a time with a 1s delay
-// https://github.com/mikedeboer/node-github/issues/526
-function rateLimitedClient (github) {
-  const limiter = new Bottleneck(1, 1000)
-  const oldHandler = github.handler
-  github.handler = (msg, block, callback) => {
-    limiter.submit(oldHandler.bind(github), msg, block, callback)
-  }
-  return github
 }
 
 module.exports = (...args) => new Robot(...args)
@@ -197,13 +176,4 @@ module.exports = (...args) => new Robot(...args)
  * A [GitHub webhook event](https://developer.github.com/webhooks/#events) payload
  *
  * @typedef payload
- */
-
-/**
- * the [github Node.js module](https://github.com/mikedeboer/node-github),
- * which wraps the [GitHub API](https://developer.github.com/v3/) and allows
- * you to do almost anything programmatically that you can do through a web
- * browser.
- * @typedef github
- * @see {@link https://github.com/mikedeboer/node-github}
  */
